+++ x-pack/solutions/observability/plugins/infra/kibana.jsonc
@@ -54,6 +54,8 @@
       "observabilityAIAssistant",
       "observabilityAgentBuilder",
       "licenseManagement",
+      "presentationPanel",
+      "unifiedChartSectionViewer",
       "serverless"
     ],
     "requiredBundles": [
